+++ fabrictestbed_extensions/fablib/slice.py
@@ -1610,7 +1610,7 @@ class Slice:
 
         self.topology = None
 
-    def renew(self, end_date: str = None, days: int = None):
+    def renew(self, end_date: str = None, days: int = None, **kwargs):
         """
         Renews the FABRIC slice's lease to the new end date.
 
@@ -1632,7 +1632,7 @@ class Slice:
             end = datetime.strptime(end_date, "%Y-%m-%d %H:%M:%S %z")
             days = (end - datetime.now(timezone.utc)).days
 
-        self.submit(lease_in_days=days, post_boot_config=False)
+        self.submit(lease_in_days=days, post_boot_config=False, **kwargs)
 
     def build_error_exception_string(self) -> str:
         """
@@ -2154,6 +2154,7 @@ class Slice:
         lease_start_time: datetime = None,
         lease_in_days: int = None,
         validate: bool = False,
+        **kwargs,
     ) -> str:
         """
         Submits a slice request to FABRIC.